+++ packages/database/prisma/migrations/20260831023412_currency_is_tradeable/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "currencies" ADD COLUMN     "is_tradeable" BOOLEAN NOT NULL DEFAULT true;